move ForgettablePayloads DI code into main EventSourcing.ForgettablePayloads package
