--- EventSourcing.ForgettablePayloads/ForgettablePayloads.Extensions.DependencyInjection/ForgettablePayloads.Extensions.DependencyInjection.csproj
@@ -1,34 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <TreatWarningsAsErrors Condition=" '$(Configuration)' == 'Release' ">true</TreatWarningsAsErrors>
-        <GenerateDocumentationFile>true</GenerateDocumentationFile>
-        <RootNamespace>EventSourcing.ForgettablePayloads.Extensions.DependencyInjection</RootNamespace>
-        <AssemblyName>EventSourcing.ForgettablePayloads.Extensions.DependencyInjection</AssemblyName>
-    </PropertyGroup>
-
-    <!-- NuGet metadata -->
-    <PropertyGroup>
-        <Title>EventSourcing.ForgettablePayloads.Extensions.DependencyInjection</Title>
-        <Description>This package contains DependencyInjection extensions for default implementation of EventSourcing.ForgettablePayloads.Abstraction package.</Description>
-        <Authors>Dawid Szmidka</Authors>
-        <PackageId>EventSourcing.ForgettablePayloads.Extensions.DependencyInjection</PackageId>
-        <Version>0.1.0-alpha</Version>
-        <PackageReleaseNotes>
-            - Initial version of EventSourcing.ForgettablePayloads.Extensions.DependencyInjection project
-        </PackageReleaseNotes>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <PackageTags>EventSourcing;EventSourcing.ForgettablePayloads.Extensions;EventSourcing.ForgettablePayloads.Extensions.DependencyInjection</PackageTags>
-        <IncludeSymbols>true</IncludeSymbols>
-        <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-        <IncludeSource>true</IncludeSource>
-        <RepositoryUrl>https://github.com/TheName/EventSourcing</RepositoryUrl>
-    </PropertyGroup>
-    
-    <ItemGroup>
-      <ProjectReference Include="..\..\Core\EventSourcing\EventSourcing.csproj" />
-      <ProjectReference Include="..\ForgettablePayloads\ForgettablePayloads.csproj" />
-    </ItemGroup>
-
-</Project>
