+++ EventSourcing.ForgettablePayloads/Persistence/ForgettablePayloads.DependencyInjection.Persistence/ForgettablePayloads.DependencyInjection.Persistence.csproj
@@ -26,7 +26,6 @@
         <RepositoryUrl>https://github.com/TheName/EventSourcing</RepositoryUrl>
     </PropertyGroup>
     <ItemGroup>
-      <ProjectReference Include="..\..\ForgettablePayloads.Extensions.DependencyInjection\ForgettablePayloads.Extensions.DependencyInjection.csproj" />
       <ProjectReference Include="..\ForgettablePayloads.Persistence\ForgettablePayloads.Persistence.csproj" />
     </ItemGroup>
