+++ EventSourcing.ForgettablePayloads/Persistence/ForgettablePayloads.DependencyInjection.Persistence/IEventSourcingForgettablePayloadsPersistenceBuilder.cs
@@ -1,4 +1,6 @@
-﻿namespace EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence
+﻿using EventSourcing.ForgettablePayloads.Abstractions.DependencyInjection;
+
+namespace EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence
 {
     /// <summary>
     /// Builder for EventSourcing.ForgettablePayloads.Persistence
