+++ EventSourcing.ForgettablePayloads/Persistence/PostgreSql/Extensions.DependencyInjection.PostgreSql.UnitTests/Extensions.DependencyInjection.PostgreSql.UnitTests.csproj
@@ -21,7 +21,9 @@
     </ItemGroup>
 
     <ItemGroup>
+        <ProjectReference Include="..\..\..\..\Core\EventSourcing\EventSourcing.csproj" />
         <ProjectReference Include="..\..\..\..\Helpers\TestHelpers\TestHelpers.csproj" />
+        <ProjectReference Include="..\..\..\ForgettablePayloads\ForgettablePayloads.csproj" />
         <ProjectReference Include="..\Extensions.DependencyInjection.PostgreSql\Extensions.DependencyInjection.PostgreSql.csproj" />
     </ItemGroup>
