+++ EventSourcing.ForgettablePayloads/Persistence/PostgreSql/Extensions.DependencyInjection.PostgreSql/EventSourcingForgettablePayloadsBuilderExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using EventSourcing.ForgettablePayloads.Abstractions.DependencyInjection;
 
 namespace EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence.PostgreSql
 {
