+++ EventSourcing.ForgettablePayloads/Persistence/PostgreSql/Extensions.DependencyInjection.PostgreSql/Extensions.DependencyInjection.PostgreSql.csproj
@@ -25,6 +25,10 @@
         <IncludeSource>true</IncludeSource>
         <RepositoryUrl>https://github.com/TheName/EventSourcing</RepositoryUrl>
     </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="5.0.0" />
+    </ItemGroup>
     
     <ItemGroup>
         <ProjectReference Include="..\..\ForgettablePayloads.DependencyInjection.Persistence\ForgettablePayloads.DependencyInjection.Persistence.csproj" />
