+++ EventSourcing.ForgettablePayloads/Persistence/SqlServer/Extensions.DependencyInjection.SqlServer.UnitTests/Extensions.DependencyInjection.SqlServer.UnitTests.csproj
@@ -21,7 +21,9 @@
     </ItemGroup>
 
     <ItemGroup>
+        <ProjectReference Include="..\..\..\..\Core\EventSourcing\EventSourcing.csproj" />
         <ProjectReference Include="..\..\..\..\Helpers\TestHelpers\TestHelpers.csproj" />
+        <ProjectReference Include="..\..\..\ForgettablePayloads\ForgettablePayloads.csproj" />
         <ProjectReference Include="..\Extensions.DependencyInjection.SqlServer\Extensions.DependencyInjection.SqlServer.csproj" />
     </ItemGroup>
