+++ EventSourcing.ForgettablePayloads/Persistence/SqlServer/Extensions.DependencyInjection.SqlServer.UnitTests/ServiceProvider_Should.cs
@@ -1,6 +1,5 @@
 ﻿using EventSourcing.Extensions;
-using EventSourcing.ForgettablePayloads.Extensions.DependencyInjection;
-using EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence;
+using EventSourcing.ForgettablePayloads.Extensions;
 using EventSourcing.ForgettablePayloads.Extensions.DependencyInjection.Persistence.SqlServer;
 using Microsoft.Extensions.DependencyInjection;
 using TestHelpers.Extensions;
@@ -10,47 +9,6 @@ namespace Extensions.DependencyInjection.SqlServer.UnitTests
 {
     public class ServiceProvider_Should
     {
-        [Fact]
-        public void BuildWithoutErrors_When_AddingEventSourcingAndExternalDependenciesAndForgettablePayloadsUsingSqlServerPersistence()
-        {
-            var serviceCollection = new ServiceCollection();
-
-            serviceCollection
-                .AddEventSourcing()
-                .WithForgettablePayloadsUsingSqlServerPersistence();
-            
-            // External dependencies
-            serviceCollection
-                .AddEventSourcingSerializationMocks()
-                .AddEventSourcingPersistenceMocks()
-                .AddEventSourcingBusMocks()
-                .AddMicrosoftLoggerMock();
-
-            serviceCollection.BuildServiceProvider(
-                new ServiceProviderOptions { ValidateScopes = true, ValidateOnBuild = true });
-        }
-        
-        [Fact]
-        public void BuildWithoutErrors_When_AddingEventSourcingAndExternalDependenciesAndForgettablePayloadsUsingSeparatePersistenceBuilder()
-        {
-            var serviceCollection = new ServiceCollection();
-
-            serviceCollection
-                .AddEventSourcing()
-                .WithForgettablePayloadsAndPersistence()
-                .WithForgettablePayloadsSqlServerPersistence();
-            
-            // External dependencies
-            serviceCollection
-                .AddEventSourcingSerializationMocks()
-                .AddEventSourcingPersistenceMocks()
-                .AddEventSourcingBusMocks()
-                .AddMicrosoftLoggerMock();
-
-            serviceCollection.BuildServiceProvider(
-                new ServiceProviderOptions { ValidateScopes = true, ValidateOnBuild = true });
-        }
-        
         [Fact]
         public void BuildWithoutErrors_When_AddingEventSourcingAndExternalDependenciesAndForgettablePayloadsUsingSeparateBuilders()
         {
