+++ EventSourcing.ForgettablePayloads/Persistence/SqlServer/Extensions.DependencyInjection.SqlServer/Extensions.DependencyInjection.SqlServer.csproj
@@ -27,6 +27,10 @@
     </PropertyGroup>
     
     <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="5.0.0" />        
+    </ItemGroup>
+    
+    <ItemGroup>
       <ProjectReference Include="..\..\ForgettablePayloads.DependencyInjection.Persistence\ForgettablePayloads.DependencyInjection.Persistence.csproj" />
       <ProjectReference Include="..\ForgettablePayloads.Persistence.SqlServer\ForgettablePayloads.Persistence.SqlServer.csproj" />
     </ItemGroup>
